moar cleanup and test fix
